+++ Projektarbeit/src/Insertionsort.java
@@ -7,7 +7,7 @@
 public class Insertionsort implements Sortieralgorithmus {
     private long vergleiche = 0;
     private double zeit;
-    private int schreibzugriffe = 0;
+    private long schreibzugriffe = 0;
     private int length;
     private int[] liste;
 
@@ -80,7 +80,7 @@ public class Insertionsort implements Sortieralgorithmus {
      * @return Gibt die Anzahl der Speicherzugriffe auf das zu sortierende Array zurück.
      */
     @Override
-    public int getSchreibzugriffe() {
+    public long getSchreibzugriffe() {
         return schreibzugriffe;
     }
 }
