+++ Projektarbeit/src/Mergesort.java
@@ -6,7 +6,7 @@
 public class Mergesort implements Sortieralgorithmus{
     private long vergleiche = 0;
     private double zeit;
-    private int schreibzugriffe = 0;
+    private long schreibzugriffe = 0;
     private int speicherbedarf;
 
     /**
@@ -136,7 +136,7 @@ public class Mergesort implements Sortieralgorithmus{
      * @return Gibt die Anzahl der Speicherzugriffe auf das zu sortierende Array zurück.
      */
     @Override
-    public int getSchreibzugriffe() {
+    public long getSchreibzugriffe() {
         return schreibzugriffe;
     }
 }
