+++ Projektarbeit/src/QuicksortFirstPivot.java
@@ -7,7 +7,7 @@
 public class QuicksortFirstPivot implements Sortieralgorithmus{
     private long vergleiche = 0;
     private double zeit;
-    private int schreibzugriffe = 0;
+    private long schreibzugriffe = 0;
     private int length;
 
     /**
@@ -104,7 +104,7 @@ public class QuicksortFirstPivot implements Sortieralgorithmus{
      * @return Gibt die Anzahl der Speicherzugriffe auf das zu sortierende Array zurück.
      */
     @Override
-    public int getSchreibzugriffe() {
+    public long getSchreibzugriffe() {
         return schreibzugriffe;
     }
 }
