+++ Projektarbeit/src/QuicksortRandomPivot.java
@@ -10,7 +10,7 @@ import java.util.Random;
 public class QuicksortRandomPivot implements Sortieralgorithmus{
     private long vergleiche = 0;
     private double zeit;
-    private int schreibzugriffe = 0;
+    private long schreibzugriffe = 0;
     private int length;
     private int[] liste;
 
@@ -119,7 +119,7 @@ public class QuicksortRandomPivot implements Sortieralgorithmus{
      * @return Gibt die Anzahl der Speicherzugriffe auf das zu sortierende Array zurück.
      */
     @Override
-    public int getSchreibzugriffe() {
+    public long getSchreibzugriffe() {
         return schreibzugriffe;
     }
